+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Double the range of fireplace and torch-like prefabs to allow counterplay to the modpack's extremely dark nights.
+
 ## [0.1.2] - 2024-04-30
 
 - Remove `Assembly.LoadFile` calls in order not to trigger Thunderstore submission rejection filter.